requirements.txt
mysql-connector-python~=9.3.0
fastapi~=0.115.12
starlette~=0.46.2
pydantic~=2.11.4